+++ changelog.md
@@ -1,6 +1,7 @@
 ##### Version 1.21.1-12.3.3-pre
 - Add a silly Metal Gear Solid reference to crates (BluSunrize)
 - Fix crashes related to the IKELOS shader on servers (BluSunrize)
+- Fix feedthrough insulator not adopting its source block's hardness (BluSunrize)
 
 ##### Version 1.21.1-12.3.2-190
 - Add support for chance-based output to the cloche (BluSunrize)